fix: move fetchTweetsNum from AppBar to router-page
